Updated per pr comments
