+++ src/test/java/org/jetbrains/research/anticopypaster/ide/RefactoringEventTest.java
@@ -22,20 +22,17 @@ import org.jetbrains.annotations.Nullable;
 import org.junit.Before;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.BeforeEach;
-// import static org.junit.Assert.assertEquals;
-import javax.swing.*;
-
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
 public class RefactoringEventTest {
-    private PsiFile file;
-    private PsiMethod destinationMethod;
-    private String text;
-    private int matches;
-    private Project project;
-    private Editor editor;
-    private int linesOfCode;
+//    private PsiFile file;
+//    private PsiMethod destinationMethod;
+//    private String text;
+//    private int matches;
+//    private Project project;
+//    private Editor editor;
+//    private int linesOfCode;
     private RefactoringEvent refactoringEvent;
 
     @BeforeEach
@@ -44,8 +41,20 @@ public class RefactoringEventTest {
     }
 
     @Test
+    public void testPrintMessage() {
+        System.out.println("Message = Test 1");
+        assertTrue(true);
+    }
+
+    @Test
+    public void testPrintMessage2() {
+        System.out.println("Message2 = Test 2");
+        assertTrue(true);
+    }
+
+    @Test
     public void testNullRefactorEvent(){
-        System.out.println("Testing boiiiii");
+        //Baseline "null" test to see that our testing folder can access src files
         assertEquals(refactoringEvent.getFile(), null);
     }
 }
