+++ src/test/resources/AntiCopyPasterBundle.properties
@@ -1,0 +1,5 @@
+extract.method.to.simplify.logic.of.enclosing.method=This code fragment can be extracted into a separate method to simplify the logic of the enclosing method.
+extract.method.to.remove.duplicated.constructor.call.or.factory.method=This code fragment can be extracted into a separate method to remove a duplicated constructor call or a factory method.
+extract.method.to.simplify.enclosing.method.and.remove.duplicates=This code fragment can be extracted into a separate method to simplify the enclosing method and remove {0} duplicates.
+extract.method.refactoring.is.available=<a href>Extract Method refactoring is available</a>
+anticopypaster.recommendation.dialog.name=AntiCopyPaster Recommendation
