--- test/main/java/org/jetbrains/research/anticopypaster/AntiCopyPasterBundleTest.java
@@ -1,4 +1,0 @@
-package main.java.org.jetbrains.research.anticopypaster;
-
-public class AntiCopyPasterBundleTest {
-}
