--- test/main/java/org/jetbrains/research/anticopypaster/checkers/FragmentCorrectnessCheckerTest.java
@@ -1,4 +1,0 @@
-package main.java.org.jetbrains.research.anticopypaster.checkers;
-
-public class FragmentCorrectnessCheckerTest {
-}
