--- test/main/java/org/jetbrains/research/anticopypaster/ide/ExtractionTaskTest.java
@@ -1,4 +1,0 @@
-package main.java.org.jetbrains.research.anticopypaster.ide;
-
-public class ExtractionTaskTest {
-}
