--- test/main/java/org/jetbrains/research/anticopypaster/ide/RefactoringEventTest.java
@@ -1,4 +1,0 @@
-package main.java.org.jetbrains.research.anticopypaster.ide;
-
-public class RefactoringEventTest {
-}
