--- test/main/java/org/jetbrains/research/anticopypaster/models/PredictionModelTest.java
@@ -1,4 +1,0 @@
-package main.java.org.jetbrains.research.anticopypaster.models;
-
-public class PredictionModelTest {
-}
