--- test/main/java/org/jetbrains/research/anticopypaster/models/ScikitModelTest.java
@@ -1,4 +1,0 @@
-package main.java.org.jetbrains.research.anticopypaster.models;
-
-public class ScikitModelTest {
-}
