--- test/main/java/org/jetbrains/research/anticopypaster/statistics/AntiCopyPasterUsageStatisticsTest.java
@@ -1,4 +1,0 @@
-package main.java.org.jetbrains.research.anticopypaster.statistics;
-
-public class AntiCopyPasterUsageStatisticsTest {
-}
