--- test/main/java/org/jetbrains/research/anticopypaster/utils/PsiUtilTest.java
@@ -1,4 +1,0 @@
-package main.java.org.jetbrains.research.anticopypaster.utils;
-
-public class PsiUtilTest {
-}
